feat(admin-ui): create redux for uma
