+++ admin-ui/plugins/auth-server/redux/actions/types.js
@@ -37,6 +37,8 @@ export const DELETE_CLIENT = 'DELETE_CLIENT'
 export const DELETE_CLIENT_RESPONSE = 'DELETE_CLIENT_RESPONSE'
 export const SET_CLIENT_ITEM = 'SET_CLIENT_ITEM'
 export const SET_VIEW = 'SET_VIEW'
+
+// UMA Resources
 export const GET_UMA_RESOURCES = 'GET_UMA_RESOURCES'
 export const GET_UMA_RESOURCES_RESPONSE = 'GET_UMA_RESOURCES_RESPONSE'
