+++ admin-ui/plugins/auth-server/redux/api/OIDCApi.js
@@ -35,14 +35,6 @@ export default class OIDCApi {
     })
   }
 
-  getUMAResources = async (clientId) => {
-    return new Promise((resolve, reject) => {
-      this.api.getOauthUmaResourcesByClientid(clientId, (error, data) => {
-        this.handleResponse(error, reject, resolve, data)
-      })
-    })
-  }
-
   handleResponse(error, reject, resolve, data) {
     if (error) {
       reject(error)
