+++ admin-ui/plugins/auth-server/redux/api/UMAResourceApi.js
@@ -1,0 +1,21 @@
+export default class UMAResourceApi {
+  constructor(api) {
+    this.api = api
+  }
+
+  getUMAResources = async (clientId) => {
+    return new Promise((resolve, reject) => {
+      this.api.getOauthUmaResourcesByClientid(clientId, (error, data) => {
+        this.handleResponse(error, reject, resolve, data)
+      })
+    })
+  }
+
+  handleResponse(error, reject, resolve, data) {
+    if (error) {
+      reject(error)
+    } else {
+      resolve(data)
+    }
+  }
+}
