+++ admin-ui/plugins/auth-server/redux/audit/Resources.js
@@ -1,3 +1,4 @@
 export const OIDC = 'openid-client'
 export const SCOPE = 'scope'
+export const UMA = 'uma'
 export const JSON_CONFIG = 'json-configuration'
