+++ admin-ui/plugins/auth-server/redux/sagas/OIDCSaga.js
@@ -8,8 +8,7 @@ import {
   getOpenidClientsResponse,
   addClientResponse,
   editClientResponse,
-  deleteClientResponse,
-  getUMAResourcesByClientResponse,
+  deleteClientResponse
 } from '../actions/OIDCActions'
 import { getAPIAccessToken } from '../actions/AuthActions'
 import { OIDC } from '../audit/Resources'
@@ -24,8 +23,7 @@ import {
   ADD_NEW_CLIENT,
   EDIT_CLIENT,
   DELETE_CLIENT,
-  SEARCH_CLIENTS,
-  GET_UMA_RESOURCES
+  SEARCH_CLIENTS
 } from '../actions/types'
 import OIDCApi from '../api/OIDCApi'
 import { getClient } from 'Redux/api/base'
@@ -42,21 +40,6 @@ function* newFunction() {
   }
   const issuer = yield select((state) => state.authReducer.issuer)
   const api = new JansConfigApi.OAuthOpenIDConnectClientsApi(
-    getClient(JansConfigApi, token, issuer),
-  )
-  return new OIDCApi(api)
-}
-
-function* newUMAFunction() {
-  const wholeToken = yield select((state) => state.authReducer.token)
-  let token = null
-  if (wholeToken) {
-    token = yield select((state) => state.authReducer.token.access_token)
-  } else {
-    token = null
-  }
-  const issuer = yield select((state) => state.authReducer.issuer)
-  const api = new JansConfigApi.OAuthUMAResourcesApi(
     getClient(JansConfigApi, token, issuer),
   )
   return new OIDCApi(api)
@@ -139,28 +122,6 @@ export function* deleteAClient({ payload }) {
   }
 }
 
-export function* getUMAResourcesByClient({ payload }) {
-  const audit = yield* initAudit()
-  try {
-    payload = payload ? payload : {}
-    addAdditionalData(audit, FETCH, GET_UMA_RESOURCES, payload)
-    const openIdApi = yield* newUMAFunction()
-    const data = yield call(
-      openIdApi.getUMAResources,
-      payload.inum,
-    )
-    yield put(getUMAResourcesByClientResponse(data))
-    yield call(postUserAction, audit)
-  } catch (e) {
-    console.log(e)
-    yield put(getUMAResourcesByClientResponse(null))
-    if (isFourZeroOneError(e)) {
-      const jwt = yield select((state) => state.authReducer.userinfo_jwt)
-      yield put(getAPIAccessToken(jwt))
-    }
-  }
-}
-
 export function* getOpenidClientsWatcher() {
   yield takeLatest(GET_OPENID_CLIENTS, getOauthOpenidClients)
 }
@@ -179,9 +140,6 @@ export function* editClientWatcher() {
 export function* deleteClientWatcher() {
   yield takeLatest(DELETE_CLIENT, deleteAClient)
 }
-export function* getUMAResourcesByClientWatcher() {
-  yield takeLatest(GET_UMA_RESOURCES, getUMAResourcesByClient)
-}
 
 export default function* rootSaga() {
   yield all([
@@ -190,6 +148,5 @@ export default function* rootSaga() {
     fork(addClientWatcher),
     fork(editClientWatcher),
     fork(deleteClientWatcher),
-    fork(getUMAResourcesByClientWatcher),
   ])
 }
